Module icu_segmenter_lstm::math_helper

Functions

change_row
change_row gets one 2d array (arr), one 1d array (arr1), and an index (row_id), and replaces the row_id-th row of arr with arr1

concatenate_arr1
concatenate_arr1 concatenates two 1d arrays to form another 1d array.

max_arr1
max_arr1 returns the index of the maximum value in a 1d array.

sigmoid
sigmoid computes the sigmoid function for a scalar value.

sigmoid_arr1
tanh_arr1 computes elementwise sigmoid funciton for elements of a 1d array

softmax
softmax gets a 1d array of f32 numbers, and compute the softmax probability for each element in the array.

tanh_arr1
tanh_arr1 computes elementwise tanh funciton for elements of a 1d array